code changes per review comments

+++ grails-app/controllers/au/org/ala/dataqualityfilter/AdminDataQualityController.groovy
@@ -14,6 +14,7 @@
  */
 package au.org.ala.dataqualityfilter
 
+import au.org.ala.web.CASRoles
 import com.google.gson.Gson
 import grails.converters.JSON
 import grails.gorm.transactions.Transactional
@@ -25,7 +26,7 @@ class AdminDataQualityController {
     public static final String DATA_PROFILES_ACTION_NAME = 'data-profiles'
     def qualityService
     def webServicesService
-    def userService
+    def authService
 
     def filters() {
         def qp = QualityProfile.get(params.long('id'))
@@ -39,27 +40,30 @@ class AdminDataQualityController {
     def 'data-profiles'() {
         def model = [:]
         // if admin, add public profiles
-        if (userService.isLoggedInAdmin()) {
+        if (request.isUserInRole(CASRoles.ROLE_ADMIN)) {
             model.put('publicProfiles', QualityProfile.findAll(sort:"id") {
                 isPublic == true
             })
         }
         // always add private profiles
         model.put('privateProfiles', QualityProfile.findAll(sort:"id") {
-            userId == userService.getLoggedInUserId()
+            userId == authService?.getUserId()
         })
-        model.put('userId', userService.getLoggedInUserId())
+        model.put('userId', authService?.getUserId())
         model.put('errors', flash.errors)
         model
     }
 
     def saveProfile(QualityProfile qualityProfile) {
+        preprocessProfile(qualityProfile)
         def invalidReq = false
         withForm {
             try {
                 qualityService.createOrUpdateProfile(qualityProfile)
             } catch (ValidationException e) {
                 flash.errors = e.errors
+            } catch (MaximumRecordNumberReachedException e) {
+                flash.message = e.message
             }
         }.invalidToken {
             invalidReq = true
@@ -84,34 +88,40 @@ class AdminDataQualityController {
     }
 
     def enableQualityProfile() {
-        withForm {
-            def qp = QualityProfile.get(params.long('id'))
-            qp.enabled = params.boolean('enabled', false)
-            qp.save(flush: true)
-        }.invalidToken {
-            // bad request
-            log.debug("ignore duplicate enable category request")
+        def qp = QualityProfile.get(params.long('id'))
+        if (processAllowed(qp)) {
+            withForm {
+                qp.enabled = params.boolean('enabled', false)
+                qp.save(flush: true)
+            }.invalidToken {
+                // bad request
+                log.debug("ignore duplicate enable category request")
+            }
+            redirect(action: DATA_PROFILES_ACTION_NAME)
         }
-        redirect(action: DATA_PROFILES_ACTION_NAME)
     }
 
     def setDefaultProfile() {
-        withForm {
-            qualityService.setDefaultProfile(params.long('id'))
-        }.invalidToken {
-            log.debug('set default profile invalid token')
+        if (processAllowed(QualityProfile.get(params.long('id')))) {
+            withForm {
+                qualityService.setDefaultProfile(params.long('id'))
+            }.invalidToken {
+                log.debug('set default profile invalid token')
+            }
+            redirect(action: DATA_PROFILES_ACTION_NAME)
         }
-        redirect(action: DATA_PROFILES_ACTION_NAME)
     }
 
     def deleteQualityProfile(QualityProfile qualityProfile) {
-        withForm {
-            qualityService.deleteProfile(qualityProfile)
-        }.invalidToken {
-            // bad request
-            log.debug("ignore duplicate delete profile request. name: {}, shortname: {}", qualityProfile.name, qualityProfile.shortName)
+        if (processAllowed(qualityProfile)) {
+            withForm {
+                qualityService.deleteProfile(qualityProfile)
+            }.invalidToken {
+                // bad request
+                log.debug("ignore duplicate delete profile request. name: {}, shortname: {}", qualityProfile.name, qualityProfile.shortName)
+            }
+            redirect(action: DATA_PROFILES_ACTION_NAME)
         }
-        redirect(action: DATA_PROFILES_ACTION_NAME)
     }
 
     def saveQualityCategory(QualityCategory qualityCategory) {
@@ -241,14 +251,16 @@ class AdminDataQualityController {
     def exportProfile() {
         QualityProfile profile = QualityProfile.get(params.long('id'))
 
-        if (profile) {
-            String fileName = 'profile_' + params.long('id')
-            response.setHeader('Content-Disposition', 'attachment; filename=' + fileName + '.json')
-            response.setContentType("text");
+        if (processAllowed(profile)) {
+            if (profile) {
+                String fileName = 'profile_' + params.long('id')
+                response.setHeader('Content-Disposition', 'attachment; filename=' + fileName + '.json')
+                response.setContentType("text");
 
-            JSON.use('exportProfile', {
-                render profile as JSON
-            })
+                JSON.use('exportProfile', {
+                    render profile as JSON
+                })
+            }
         }
     }
 
@@ -270,6 +282,8 @@ class AdminDataQualityController {
                 flash.message = e.getLocalizedMessage()
                 break
             }
+
+            preprocessProfile(profile)
 
             try {
                 // set profile display order, there may already have existing profiles, so new display order = current max + 1
@@ -300,11 +314,12 @@ class AdminDataQualityController {
                     break;
                 }
 
-                profile.userId = request.getParameter('userId') ?: null
                 // safe whole profile, if any filed fails validation an exception will be thrown
                 qualityService.createOrUpdateProfile(profile)
             } catch (ValidationException e) {
                 flash.errors = e.errors
+            } catch (MaximumRecordNumberReachedException e) {
+                flash.message = e.message
             }
             break
         }
@@ -321,5 +336,29 @@ class AdminDataQualityController {
         } else {
             render status: 404
         }
+    }
+
+    private boolean processAllowed(qualityProfile) {
+        while (true) {
+            // if processing public profile, you need to be an admin
+            if (!qualityProfile.userId) {
+                if (!request.isUserInRole(CASRoles.ROLE_ADMIN)) {
+                    flash.message = 'you need to be an administrator to edit/export public profiles'
+                    break;
+                }
+            } else { // if processing private profile, you need to be the owner of the profile
+                if (qualityProfile.userId != authService?.getUserId()) {
+                    flash.message = 'you can only edit/export your own profiles'
+                    break;
+                }
+            }
+
+            return true
+        }
+        redirect(controller: "adminDataQuality", action: "data-profiles")
+    }
+
+    private void preprocessProfile(qualityProfile) {
+        qualityProfile.userId = params.isPublicProfile == 'true' ? null : authService?.getUserId()
     }
 }
